Feat(customers): corrige nome dos arquivos
