+++ app/Http/Requests/Costomer/StoreCostomerRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Customer;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreCostumerRequest extends FormRequest
+class StoreCustomerRequest extends FormRequest
 {
     public function rules()
     {
