+++ routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
 use App\Http\Controllers\Car\Cars;
-use App\Http\Controllers\Customer\Customers;
+use App\Http\Controllers\Customer\CustomerController;
 use App\Http\Controllers\Rent\RentCars;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
@@ -17,7 +17,7 @@ Route::get('/home', 'HomeController@index')->name('home');
 Route::namespace('\\')
     ->middleware('auth')
     ->group(function () {
-        Route::resource('customers', Customers::class);
+        Route::resource('customers', CustomerController::class);
     });
 
 Route::namespace('\\')
